+++ SIP/SIPUtility.cpp
@@ -99,7 +99,7 @@ void SIP::make_branch( char * branch )
 	uint64_t r1 = random();
 	uint64_t r2 = random();
 	uint64_t val = (r1<<32) + r2;
-	sprintf(branch,"z9hG4bKobts28%llx", val);
+	sprintf(branch,"z9hG4bKobts28%llx", static_cast<long long unsigned int>(val));
 }
 
 /* get the return address from the SIP VIA header
